Set up TodoFooter React class
This commit has several important changes:
1. A `TodoFooter` React class was created, which includes both a count
of remaining todos as well as a function to complete all todos.
2. In order to properly update the count whenever a todo is completed
(or uncompleted), the parent TodoBox's state has to update. This
requires a method TodoBox which is invoked from the completed Todo item.
3. Additionally, Todo items weren't updating in the DOM when the "Mark
all as complete" link was clicked. This was because the `completed`
attribute on a Todo item is passed down from `TodoBox.data` as a
property, which doesn't invoke a state change. The solution to this is
to override the `componentWillReceiveProps` method and manually set the
state, which will re-render the Todo component if the state has changed.

Some other small changes in this commit include styling for the
TodoFooter and a small HTML fix to include underscore.min.js.